Add files via upload
As you can see in this unzipped file of the snakes and ladders JavaFx OOP GUI project all the files in here are mostly the images used for the project, whereas the code itself lies within the SnakesandLadders.java, the Cell.java, and the Board.java. The Board.java essentially has the necessary details in order to create the board, the cell class has the details in regards to the single cell. The snakesandladders.java has all the necessary additions to create the whole game. Again, this is a 2 player GUI snakes and ladders game created using JavaFX.